+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Common & useful classes, resources, extensions. Based on Symfony framework.
 # 0.1.1
 
 1. Add descriptor of application
+2. Replenish documentation > Descriptor of bundle + Descriptor of application
 
 # 0.1.0