+++ CHANGELOG.md
@@ -5,7 +5,10 @@ Common & useful classes, resources, extensions. Based on Symfony framework.
 # 0.1.1
 
 1. Add descriptor of application
-2. Replenish documentation > Descriptor of bundle + Descriptor of application
+2. Replenish documentation:
+	- Descriptor of bundle
+	- Descriptor of application
+	- Services: ApplicationService
 
 # 0.1.0